Sort changes into file trees and print them
The idea is to print all changes from an addition and a removal standpoint.
The addition shows where stuff was moved from and the removal shows where it was moved to.
Both contain different information - thus (for now) they are printed separately.

This commit will also make a hard-fail if the Git is unkept and the same Unity Guid was used more than once - the code cannot handle that.
This is to keep the code complexity down during the development phase.